extract permbot-stub helpers + align stderr prefix
startPermbotStub, makeSock, and captureIRC were copied across 4 test
files. Extract to test/helpers/permbot-stub.ts. captureIRC consolidates
the two slightly different buffer-management implementations, using the
cleaner lines.pop() form from irc-permission-prompt.test.ts.

permission-prompt.ts was logging perm-hook: while pretooluse-prompt.ts
logged pretooluse-hook[${WORKER}]:. Align to include [WORKER] so
operators can grep per-worker log lines from both hooks.

## Human-Claude Interaction Log

### Human prompts (VERBATIM - include typos, informal language, COMPLETE text):
1. "/worker roost 281 AvesAlight/roost maint/281-test-helpers-extract alex"
   → Claude: Read issue #281, joined #roost-issue-281, posted implementation plan

2. (roost-lead-pm via IRC) "approved. the prefix interpretation is right — keep the per-hook discriminator (perm-hook vs pretouluse-hook) so an operator grepping logs can tell them apart, just align the [${WORKER}]: format. go."
   → Claude: Implemented the extraction and prefix alignment
