+++ test/pretooluse-prompt.test.ts
@@ -5,6 +5,7 @@ import * as os from 'node:os'
 import * as path from 'node:path'
 import { classifyBash } from '../src/pretooluse-prompt.js'
 import { suppressLateRejection } from './helpers/tool.js'
+import { startPermbotStub, makeSock } from './helpers/permbot-stub.js'
 
 const HOOK = path.join(import.meta.dirname, '../src/pretooluse-prompt.ts')
 
@@ -144,32 +145,6 @@ describe('classifyBash', () => {
 })
 
 // ---- Hook subprocess --------------------------------------------------------
-
-/** Minimal permbot socket stub: reads one JSON line, responds immediately.
- *  Returns a ready promise (server is listening) and a done promise (response sent). */
-function startPermbotStub(sockPath: string, reply: object): { ready: Promise<void>; done: Promise<void> } {
-  let onReady!: () => void
-  const ready = new Promise<void>(r => { onReady = r })
-  let onDone!: () => void
-  const done = new Promise<void>(r => { onDone = r })
-  const server = net.createServer((sock) => {
-    let buf = ''
-    sock.on('data', (d) => {
-      buf += d.toString('utf8')
-      if (!buf.includes('\n')) return
-      sock.write(JSON.stringify(reply) + '\n')
-      sock.end()
-      server.close()
-      onDone()
-    })
-  })
-  server.listen(sockPath, () => { onReady() })
-  return { ready, done }
-}
-
-function makeSock(): string {
-  return path.join(os.tmpdir(), `pretooluse-hook-test-${process.pid}-${Math.random().toString(36).slice(2)}.sock`)
-}
 
 const SAFETY_CHECK_PAYLOAD = JSON.stringify({
   tool_name: 'Bash',
